Dzaddyssecret/reservation/serializers.py
from dataclasses import fields
from rest_framework import serializers
from .models import Reservation


class ReservationSerializer(serializers.ModelSerializer):
    class Meta:
        model=Reservation
        fields='__all__' 

