Dzaddyssecret/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.19
botocore==1.27.19
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.10
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
